+++ src/core/create_manifest_data.ts
@@ -1,6 +1,5 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import svelte from 'svelte/compiler';
 import { Page, PageComponent, ServerRoute, ManifestData } from '../interfaces';
 import { posixify, reserved_words } from '../utils';
 
@@ -13,28 +12,14 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 		throw new Error(`As of Sapper 0.21, the routes/ directory should become src/routes/`);
 	}
 
-	function has_preload(file: string) {
-		const source = fs.readFileSync(path.join(cwd, file), 'utf-8');
-
-		if (/preload/.test(source)) {
-			try {
-				const { vars } = svelte.compile(source.replace(/<style\b[^>]*>[^]*?<\/style>/g, ''), { generate: false });
-				return vars.some((variable: any) => variable.module && variable.export_name === 'preload');
-			} catch (err) {}
-		}
-
-		return false;
-	}
-
 	function find_layout(file_name: string, component_name: string, dir: string = '') {
 		const ext = component_extensions.find((ext) => fs.existsSync(path.join(cwd, dir, `${file_name}${ext}`)));
-		const file = posixify(path.join(dir, `${file_name}${ext}`))
+		const file = posixify(path.join(dir, `${file_name}${ext}`));
 
 		return ext
 			? {
 				name: component_name,
-				file: file,
-				has_preload: has_preload(file)
+				file
 			}
 			: null;
 	}
@@ -47,16 +32,14 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 		default: true,
 		type: 'layout',
 		name: '_default_layout',
-		file: null,
-		has_preload: false
+		file: null
 	};
 
 	const default_error: PageComponent = {
 		default: true,
 		type: 'error',
 		name: '_default_error',
-		file: null,
-		has_preload: false
+		file: null
 	};
 
 	function walk(
@@ -64,8 +47,8 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 		parent_segments: Part[][],
 		parent_params: string[],
 		stack: Array<{
-			component: PageComponent,
-			params: string[]
+			component: PageComponent;
+			params: string[];
 		}>
 	) {
 		const items = fs.readdirSync(dir)
@@ -94,7 +77,7 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 				const route_suffix = basename.slice(basename.indexOf('.'), -ext.length);
 
 				parts.forEach(part => {
-					if (part.qualifier && /[\(\)\?\:]/.test(part.qualifier.slice(1, -1))) {
+					if (part.qualifier && /[()?:]/.test(part.qualifier.slice(1, -1))) {
 						throw new Error(`Invalid route ${file} — cannot use (, ), ? or : in route qualifiers`);
 					}
 				});
@@ -161,15 +144,14 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 			else if (item.is_page) {
 				const component = {
 					name: get_slug(item.file),
-					file: item.file,
-					has_preload: has_preload(item.file)
+					file: item.file
 				};
 
 				components.push(component);
 
 				const parts = (item.is_index && stack[stack.length - 1] === null)
 					? stack.slice(0, -1).concat({ component, params })
-					: stack.concat({ component, params })
+					: stack.concat({ component, params });
 
 				pages.push({
 					pattern: get_pattern(segments, true),
@@ -182,7 +164,7 @@ export default function create_manifest_data(cwd: string, extensions: string = '
 					name: `route_${get_slug(item.file)}`,
 					pattern: get_pattern(segments, !item.route_suffix),
 					file: item.file,
-					params: params
+					params
 				});
 			}
 		});
@@ -237,12 +219,12 @@ type Part = {
 
 function is_spread(path: string) {
 	const spread_pattern = /\[\.{3}/g;
-	return spread_pattern.test(path)
+	return spread_pattern.test(path);
 }
 
 function comparator(
-	a: { basename: string, parts: Part[], file: string, is_index: boolean },
-	b: { basename: string, parts: Part[], file: string, is_index: boolean }
+	a: { basename: string; parts: Part[]; file: string; is_index: boolean; },
+	b: { basename: string; parts: Part[]; file: string; is_index: boolean; }
 ) {
 	if (a.is_index !== b.is_index) {
 		if (a.is_index) return is_spread(a.file) ? 1 : -1;
@@ -261,7 +243,7 @@ function comparator(
 
 		// if spread && index, order later
 		if (a_sub_part.spread && b_sub_part.spread) {
-			return a.is_index ? 1 : -1
+			return a.is_index ? 1 : -1;
 		}
 
 		// If one is ...spread order it later
@@ -305,7 +287,7 @@ function get_parts(part: string): Part[] {
 
 			const [, content, qualifier] = dynamic
 				? /([^(]+)(\(.+\))?$/.exec(str)
-				: [, str, null];
+				: [null, str, null];
 
 			return {
 				content,
@@ -319,12 +301,12 @@ function get_parts(part: string): Part[] {
 
 function get_slug(file: string) {
 	let name = file
-		.replace(/[\\\/]index/, '')
-		.replace(/[\/\\]/g, '_')
+		.replace(/[\\/]index/, '')
+		.replace(/[/\\]/g, '_')
 		.replace(/\.\w+$/, '')
 		.replace(/\[([^(]+)(?:\([^(]+\))?\]/, '$$$1')
 		.replace(/[^a-zA-Z0-9_$]/g, c => {
-			return c === '.' ? '_' : `$${c.charCodeAt(0)}`
+			return c === '.' ? '_' : `$${c.charCodeAt(0)}`;
 		});
 
 	if (reserved_words.has(name)) name += '_';
@@ -335,7 +317,7 @@ function get_pattern(segments: Part[][], add_trailing_slash: boolean) {
 	const path = segments.map(segment => {
 		return segment.map(part => {
 			return part.dynamic
-				? part.qualifier || (part.spread ? '(.+)' : '([^\\/]+?)')
+				? part.qualifier || (part.spread ? '(.+)' : '([^/]+?)')
 				: encodeURI(part.content.normalize())
 					.replace(/\?/g, '%3F')
 					.replace(/#/g, '%23')