+++ test/apps/errors/rollup.config.js
@@ -1,5 +1,5 @@
-import resolve from 'rollup-plugin-node-resolve';
-import replace from 'rollup-plugin-replace';
+import resolve from '@rollup/plugin-node-resolve';
+import replace from '@rollup/plugin-replace';
 import svelte from 'rollup-plugin-svelte';
 
 const mode = process.env.NODE_ENV;
@@ -27,7 +27,7 @@ export default {
 
 	server: {
 		input: config.server.input(),
-		output: config.server.output(),
+		output: Object.assign({}, config.server.output(), { sourcemap: true }),
 		plugins: [
 			replace({
 				'process.browser': false,