+++ test/apps/with-sourcemaps-webpack/webpack.config.js
@@ -33,7 +33,7 @@ module.exports = {
 			new webpack.DefinePlugin({
 				'process.browser': true,
 				'process.env.NODE_ENV': JSON.stringify(mode)
-			}),
+			})
 		].filter(Boolean),
 		devtool: dev ? 'inline-source-map' : 'source-map'
 	},